Improve math and simulation for deeper runs

+++ src/app.rs
@@ -5,8 +5,8 @@ use crate::generators::{pattern_by_name, random_soup};
 pub fn initial_grid(config: &Config) -> BitGrid {
     if config.pattern == "random" {
         return random_soup(
-            (config.width as i32) * 2 / 3,
-            config.height as i32,
+            ((config.width as i64) * 2) / 3,
+            config.height as i64,
             37,
             config.seed,
         );
